Some minor fixes indicated by Solaris cc warnings.  Removed some bad (void*) casts to the fourth (function) arg of qsort;  changed game.randseed to type int instead of RANDOM_TYPE (= uint), so the (int*) in settings can point to it properly, and because we can't store a non-int value in it anyway; fixed some trivial cases of "statement not reached".
[[originally from CVS; cvs2svn created svn r850]]

+++ server/civserver.c
@@ -319,7 +319,7 @@ int main(int argc, char *argv[])
   if(!game.randseed) {
     /* We strip the high bit for now because neither game file nor
        server options can handle unsigned ints yet. - Cedric */
-    game.randseed = (unsigned int) time(NULL) & (MAX_UINT32 >> 1);
+    game.randseed = time(NULL) & (MAX_UINT32 >> 1);
   }
  
   if(!rand_init)
